forcing cast to prevent narrowing warning

+++ cocos/platform/win32/CCDevice.cpp
@@ -40,7 +40,7 @@ int Device::getDPI()
         HDC hScreenDC = GetDC( nullptr );
         int PixelsX = GetDeviceCaps( hScreenDC, HORZRES );
         int MMX = GetDeviceCaps( hScreenDC, HORZSIZE );
-        ReleaseDC( nullptr, hScreenDC );   
+        ReleaseDC( nullptr, hScreenDC );
         dpi = 254.0f*PixelsX/MMX/10;
     }
     return dpi;
@@ -80,7 +80,7 @@ public:
     wchar_t * utf8ToUtf16(const std::string& str)
     {
         wchar_t * pwszBuffer = nullptr;
-        do 
+        do
         {
             if (str.empty())
             {
@@ -88,13 +88,13 @@ public:
             }
             // utf-8 to utf-16
             int nLen = str.size();
-            int nBufLen  = nLen + 1;			
+            int nBufLen  = nLen + 1;
             pwszBuffer = new wchar_t[nBufLen];
             CC_BREAK_IF(! pwszBuffer);
             memset(pwszBuffer,0,nBufLen);
-            nLen = MultiByteToWideChar(CP_UTF8, 0, str.c_str(), nLen, pwszBuffer, nBufLen);		
+            nLen = MultiByteToWideChar(CP_UTF8, 0, str.c_str(), nLen, pwszBuffer, nBufLen);
             pwszBuffer[nLen] = '\0';
-        } while (0);	
+        } while (0);
         return pwszBuffer;
 
     }
@@ -102,7 +102,7 @@ public:
     bool setFont(const char * pFontName = nullptr, int nSize = 0)
     {
         bool bRet = false;
-        do 
+        do
         {
             std::string fontName = pFontName;
             std::string fontPath;
@@ -111,7 +111,7 @@ public:
             LOGFONTA    tOldFont = {0};
             GetObjectA(hDefFont, sizeof(tNewFont), &tNewFont);
             if (fontName.c_str())
-            {    
+            {
                 // create font from ttf file
                 int nFindttf = fontName.find(".ttf");
                 int nFindTTF = fontName.find(".TTF");
@@ -121,7 +121,7 @@ public:
                     int nFindPos = fontName.rfind("/");
                     fontName = &fontName[nFindPos+1];
                     nFindPos = fontName.rfind(".");
-                    fontName = fontName.substr(0,nFindPos);                
+                    fontName = fontName.substr(0,nFindPos);
                 }
                 else
                 {
@@ -131,7 +131,7 @@ public:
                         if (fontName.length() == nFindPos + 1)
                         {
                             fontName = "";
-                        } 
+                        }
                         else
                         {
                             fontName = &fontName[nFindPos+1];
@@ -150,7 +150,7 @@ public:
             if (tOldFont.lfHeight == tNewFont.lfHeight
                 && 0 == strcmp(tOldFont.lfFaceName, tNewFont.lfFaceName))
             {
-                // already has the font 
+                // already has the font
                 bRet = true;
                 break;
             }
@@ -167,7 +167,7 @@ public:
                     if(AddFontResource(pwszBuffer))
                     {
                         SendMessage( _wnd, WM_FONTCHANGE, 0, 0);
-                    }						
+                    }
                     delete [] pwszBuffer;
                     pwszBuffer = nullptr;
                 }
@@ -195,7 +195,7 @@ public:
     SIZE sizeWithText(const wchar_t * pszText, int nLen, DWORD dwFmt, LONG nWidthLimit)
     {
         SIZE tRet = {0};
-        do 
+        do
         {
             CC_BREAK_IF(! pszText || nLen <= 0);
 
@@ -246,7 +246,7 @@ public:
     {
         int nRet = 0;
         wchar_t * pwszBuffer = 0;
-        do 
+        do
         {
             CC_BREAK_IF(! pszText);
 
@@ -362,7 +362,7 @@ private:
     HWND    _wnd;
     std::string _curFontPath;
 
-    void removeCustomFont() 
+    void removeCustomFont()
     {
         HFONT hDefFont = (HFONT)GetStockObject(DEFAULT_GUI_FONT);
         if (hDefFont != _font)
@@ -370,7 +370,7 @@ private:
             DeleteObject(_font);
             _font = hDefFont;
         }
-        // release temp font resource	
+        // release temp font resource
         if (_curFontPath.size() > 0)
         {
             wchar_t * pwszBuffer = utf8ToUtf16(_curFontPath);
@@ -382,7 +382,7 @@ private:
                 pwszBuffer = nullptr;
             }
             _curFontPath.clear();
-        }	
+        }
     }
 };
 
@@ -395,7 +395,7 @@ static BitmapDC& sharedBitmapDC()
 Data Device::getTextureDataForText(const char * text, const FontDefinition& textDefinition, TextAlign align, int &width, int &height, bool& hasPremultipliedAlpha)
 {
     Data ret;
-    do 
+    do
     {
         BitmapDC& dc = sharedBitmapDC();
 
@@ -405,7 +405,7 @@ Data Device::getTextureDataForText(const char * text, const FontDefinition& text
         }
 
         // draw text
-        SIZE size = {textDefinition._dimensions.width, textDefinition._dimensions.height};
+        SIZE size = {(long) textDefinition._dimensions.width, (long) textDefinition._dimensions.height};
         CC_BREAK_IF(! dc.drawText(text, size, align));
 
         int dataLen = size.cx * size.cy * 4;
@@ -418,7 +418,7 @@ Data Device::getTextureDataForText(const char * text, const FontDefinition& text
             int mask[4];
         } bi = {0};
         bi.bmiHeader.biSize = sizeof(bi.bmiHeader);
-        CC_BREAK_IF(! GetDIBits(dc.getDC(), dc.getBitmap(), 0, 0, 
+        CC_BREAK_IF(! GetDIBits(dc.getDC(), dc.getBitmap(), 0, 0,
             nullptr, (LPBITMAPINFO)&bi, DIB_RGB_COLORS));
 
         width    = (short)size.cx;
@@ -427,7 +427,7 @@ Data Device::getTextureDataForText(const char * text, const FontDefinition& text
         // copy pixed data
         bi.bmiHeader.biHeight = (bi.bmiHeader.biHeight > 0)
             ? - bi.bmiHeader.biHeight : bi.bmiHeader.biHeight;
-        GetDIBits(dc.getDC(), dc.getBitmap(), 0, height, dataBuf, 
+        GetDIBits(dc.getDC(), dc.getBitmap(), 0, height, dataBuf,
             (LPBITMAPINFO)&bi, DIB_RGB_COLORS);
 
         // change pixel's alpha value to 255, when it's RGB != 0
